+++ docs/changelog.rst
@@ -10,7 +10,7 @@ Unreleased
 
 Added
 ^^^^^
-- add options to calculate daily Penman-Monteith potential evaporation using the pyeto package. Depending on the available variables, two options are defined ``penman-monteith_era5`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind_u', 'wind_v', 'rh', 'kin', 'press_msl']) and ``penman-monteith_nasanex`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind', 'rh', 'kin']).
+- add options to calculate daily Penman-Monteith potential evaporation using the pyeto package. Depending on the available variables, two options are defined ``penman-monteith_tdew`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind_u', 'wind_v', 'temp_dew', 'kin', 'press_msl']) and ``penman-monteith_rh_simple`` (inputs: ['temp', 'temp_min', 'temp_max', 'wind', 'rh', 'kin']).
 
 Fixed
 ^^^^^